+++ modules/sasiae/include/device/eirbot2014/motor.hpp
