+++ modules/sasiae/include/device/output_device.hpp
